pipecat-voiceagent/requirements_rustpbx.txt
python-dotenv
loguru
websockets>=12.0
pipecat-ai[cartesia,openai,silero]>=0.0.82
